Fixes tests
